move around files for a more consistent scanner file organization
