+++ app/components/scanner/code-scanner.tsx
