+++ app/components/scanner/success-animation.tsx
