+++ app/components/scanner/utils.tsx
